+++ package/cosmos_foundation/CHANGELOG.md
@@ -22,6 +22,8 @@
 * Added the Responsive helper that provides powerful methods to calculate dynamic response properties.
   * With a method to calculate traditional breakpoints as small, medium, and large devices.
   * With a mehtod to calculate powerful dynamic breakpoints and values during runtime.  
+  * Added a shortcut to indicate if the current window is larger.
+  * Small improvements in the internal objects management.
 * Improved theme helper documentation and API communication.
   * Added an updateEffect to getTheme method to notify all stateful widgets listeners about
   the update of the current theme to use.